New description for the game
Ah, I was a little mistaken and inserted the description in the wrong commit) Well, this is my first experience using Git, so do not scold it strictly))
I wanted to attach the description of the second commit to the first, and the changes that I made to the second commit were simple code formatting and structure optimization.
This time I added a description for the game in Game.h.